ui: Replace specific components with generic upload/download
