+++ ui/src/components/ArtifactFiles.tsx
@@ -1,13 +1,21 @@
 import { Table } from "react-bootstrap";
-import { useQuery } from "@tanstack/react-query";
+import { useQuery, useQueryClient } from "@tanstack/react-query";
+import { BsDownload } from "react-icons/bs";
 import { useParams } from "react-router";
-import { fetchArtifact } from "api/artifacts";
+import {
+  downloadArtifactFile,
+  fetchArtifact,
+  uploadArtifactFile,
+} from "api/artifacts";
 import ArtifactFileDeleteButton from "components/ArtifactFileDeleteButton";
-import ArtifactFileDownloadButton from "components/ArtifactFileDownloadButton";
-import ArtifactFileUploader from "components/ArtifactFileUploader";
+import DownloadButton from "components/DownloadButton";
+import FileUploader from "components/FileUploader";
+import { useNotification } from "context/notificationContext";
 
 const ArtifactFiles = () => {
   const { uuid } = useParams();
+  const queryClient = useQueryClient();
+  const { notify } = useNotification();
 
   const {
     data: artifact = null,
@@ -17,6 +25,29 @@ const ArtifactFiles = () => {
     queryKey: ["artifacts", uuid],
     queryFn: () => fetchArtifact(uuid),
   });
+
+  const handleDownload = async (artifactUUID: string, filename: string) => {
+    return await downloadArtifactFile(artifactUUID, filename);
+  };
+
+  const handleUpload = async (file: File | null) => {
+    return await uploadArtifactFile(uuid, file)
+      .then((response) => {
+        if (response.error_code == 0) {
+          notify.success(`Artifact file uploaded`);
+          void queryClient.invalidateQueries({
+            queryKey: ["artifacts", uuid],
+          });
+          return true;
+        }
+        notify.error(response.error);
+        return false;
+      })
+      .catch((e) => {
+        notify.error(`Error during artifact creation: ${e}`);
+        return false;
+      });
+  };
 
   if (isLoading) {
     return <div>Loading...</div>;
@@ -37,9 +68,14 @@ const ArtifactFiles = () => {
                   {item}
                 </td>
                 <td className="w-75">
-                  <ArtifactFileDownloadButton
-                    artifactUUID={uuid || ""}
-                    fileName={item}
+                  <DownloadButton
+                    title="Download"
+                    size="sm"
+                    variant="outline-secondary"
+                    className="bg-white border no-hover m-2"
+                    onDownload={() => handleDownload(uuid || "", item)}
+                    filename={item}
+                    children={<BsDownload />}
                   />
                   <ArtifactFileDeleteButton
                     artifactUUID={uuid || ""}
@@ -52,7 +88,7 @@ const ArtifactFiles = () => {
         </tbody>
       </Table>
       <div>
-        <ArtifactFileUploader uuid={uuid} />
+        <FileUploader onUpload={handleUpload} />
       </div>
     </div>
   );
